+++ aquadoggo/src/graphql/client/query.rs
@@ -6,12 +6,10 @@ use p2panda_rs::identity::Author;
 use p2panda_rs::storage_provider::traits::StorageProvider;
 
 use crate::db::provider::SqlStorage;
-use crate::db::Pool;
+use crate::graphql::client::{EntryArgsRequest, EntryArgsResponse};
 
-use super::{EntryArgsRequest, EntryArgsResponse};
-
-#[derive(Default, Debug, Copy, Clone)]
 /// The GraphQL root for the client api that p2panda clients can use to connect to a node.
+#[derive(Default, Debug, Copy, Clone)]
 pub struct Query;
 
 #[Object]
@@ -41,13 +39,9 @@ impl Query {
             document: document_id,
         };
 
-        // Prepare database connection
-        let pool = ctx.data::<Pool>()?;
-        let provider = SqlStorage {
-            pool: pool.to_owned(),
-        };
-
-        provider.get_entry_args(&args).await.map_err(Error::from)
+        // Load and return next entry arguments
+        let store = ctx.data::<SqlStorage>()?;
+        store.get_entry_args(&args).await.map_err(Error::from)
     }
 }
 
@@ -61,12 +55,12 @@ mod tests {
     use crate::context::Context;
     use crate::graphql::client::EntryArgsResponse;
     use crate::server::build_server;
-    use crate::test_helpers::{initialize_db, TestClient};
+    use crate::test_helpers::{initialize_store, TestClient};
 
     #[tokio::test]
     async fn next_entry_args_valid_query() {
-        let pool = initialize_db().await;
-        let context = Context::new(pool.clone(), Configuration::default());
+        let store = initialize_store().await;
+        let context = Context::new(store, Configuration::default());
         let client = TestClient::new(build_server(context));
 
         // Selected fields need to be alphabetically sorted because that's what the `json` macro
@@ -87,7 +81,6 @@ mod tests {
             }))
             .send()
             .await
-            // .json::<GQLResponse<EntryArgsGQLResponse>>()
             .json::<Response>()
             .await;
 
@@ -110,13 +103,12 @@ mod tests {
 
     #[tokio::test]
     async fn next_entry_args_error_response() {
-        let pool = initialize_db().await;
-        let context = Context::new(pool.clone(), Configuration::default());
+        let store = initialize_store().await;
+        let context = Context::new(store, Configuration::default());
         let client = TestClient::new(build_server(context));
 
         // Selected fields need to be alphabetically sorted because that's what the `json` macro
         // that is used in the assert below produces.
-
         let response = client
             .post("/graphql")
             .json(&json!({