+++ aquadoggo/src/graphql/schema.rs
@@ -2,17 +2,17 @@
 
 use async_graphql::{EmptySubscription, Schema};
 
-use crate::db::Pool;
+use crate::db::provider::SqlStorage;
 use crate::graphql::client::{Mutation, Query};
 
 /// GraphQL schema for p2panda node.
 pub type RootSchema = Schema<Query, Mutation, EmptySubscription>;
 
-pub fn build_root_schema(pool: Pool) -> RootSchema {
+pub fn build_root_schema(store: SqlStorage) -> RootSchema {
     let query = Query::default();
     let mutation = Mutation::default();
 
     Schema::build(query, mutation, EmptySubscription)
-        .data(pool)
+        .data(store)
         .finish()
 }